requirements.txt
requests~=2.32.3
setuptools~=68.2.0
PyYAML~=6.0.2